+++ pkg/scheduler/scheduler_test.go
@@ -45,7 +45,7 @@ func TestMain(m *testing.M) {
 }
 
 func setupScheduler(t *testing.T, reg prometheus.Registerer) (*Scheduler, schedulerpb.SchedulerForFrontendClient, schedulerpb.SchedulerForQuerierClient) {
-	cfg := Config{}
+	cfg := Config{AdditionalQueryQueueDimensionsEnabled: true}
 	flagext.DefaultValues(&cfg)
 	cfg.MaxOutstandingPerTenant = testMaxOutstandingPerTenant
 
@@ -306,7 +306,7 @@ func TestTracingContext(t *testing.T) {
 	require.Equal(t, 1, len(scheduler.pendingRequests))
 
 	for _, r := range scheduler.pendingRequests {
-		require.NotNil(t, r.parentSpanContext)
+		require.NotNil(t, r.ParentSpanContext)
 	}
 }